refactor: backup command structure and improve file handling
- Reorganized command classes into separate modules for better maintainability.
- Updated BackupCommand to use pathlib for file path handling.
- Enhanced CleanupCommand to manage old backup files with improved logging.
- Refined EncryptCommand and DecryptCommand for better security practices and error handling.
- Introduced ExtractCommand for safe extraction of backup archives with path traversal checks.
- Updated BackupConfig to utilize pathlib for path management and improved configuration loading.
- Enhanced SizeCalculator to work with pathlib and improve directory size calculations.
- Improved security context management in ContextManager for password handling.
- Added backward compatibility for command imports in commands.py.
