--- requirements.txt
@@ -1,2 +1,0 @@
-pip==24.3.1
-tqdm==4.67.0
