+++ src/commands/__init__.py
@@ -1,0 +1,20 @@
+"""Command pattern implementations for backup operations.
+
+This module aggregates all command classes for easy importing.
+"""
+
+from src.commands.command import Command
+from src.commands.backup_command import BackupCommand
+from src.commands.cleanup_command import CleanupCommand
+from src.commands.decrypt_command import DecryptCommand
+from src.commands.encrypt_command import EncryptCommand
+from src.commands.extract_command import ExtractCommand
+
+__all__ = [
+    "Command",
+    "BackupCommand",
+    "CleanupCommand",
+    "DecryptCommand",
+    "EncryptCommand",
+    "ExtractCommand",
+]
